==> code/01_Data_Preprocess.ipynb <==
#!pip install nltk

import pandas as pd
import pyarrow as pa
import pyarrow.parquet as pq
import re
from tqdm import tqdm
import matplotlib.pyplot as plt
import os, requests, sys

import nltk as nltk
import nltk.corpus  
from nltk.text import Text
from nltk.corpus import stopwords
from nltk.stem.wordnet import WordNetLemmatizer

%%time

df = pd.read_parquet('https://storage.googleapis.com/msca-bdp-data-open/news_final_project/news_final_project.parquet', engine='pyarrow')
df.shape

df.head()

df['language'].value_counts()

df.drop(columns = ['language'], inplace = True)

### Check duplicate rows

df[df.duplicated(subset = ['title', 'text'], keep = False)].sort_values(by = 'title')

df.drop_duplicates(subset = ['title', 'text'], inplace = True)

df.shape

### Extract the main text body

def text_body(text):
    
    # split the paragraphs into lists of sentences 
    paragraphs = text.split('\n')
    
    # discard sentences after these key words, paragraphs ahead of these words can be viewed as the main text body
    exclude = ['Articles:', 'You May Also Like', 'Related Stories', 'Related Articles', 
               'Contact Us:', 'Exclusive:', 'More Stories', 'About Us:', 'PRESS RELEASE:',
               'Top 10 News', 'More Articles', 'Terms of Use', 'Privacy Policy']
    idx = 999999999

    for term in exclude:

        for index, sentence in enumerate(paragraphs):
            if term in sentence and index < idx:
                idx = index
    
    try:
        paragraphs = paragraphs[:idx]
    except:
        pass 
    
    # discard sentences that have less than 15 words
    sent = [sentence for sentence in paragraphs if len(sentence.strip().split(' ')) > 15]
    
    res = '\n'.join(sent)
    
    # remove \n and \t
    res = res.replace('\n', ' ').replace('\t', '')
    
    ### eliminate urls
    url_pattern1 = 'https://[\w\./]+' #starts with https:
    url_pattern2 = 'www.[\w\./]+' #starts with www.
    res = re.sub(url_pattern1, '', res)
    res = re.sub(url_pattern2, '', res)
    
    ### remove all right reserved
    res = re.sub('All Right Reserved', '', res)
    
    return res

%%time
df['main_text'] = df['text'].apply(text_body)

### Discard irrelevant news articles

def relevant(text):
    
    relevant = 0
    
    topics = ['Data Science', ' DS ', 'Machine Learning', 'Artificial Intelligence', ' AI ', 'data science', 'machine learning']
    
    for topic in topics:
        
        if topic in text: relevant = 1
        
    return relevant

df['relevant'] = df['main_text'].apply(relevant)

filtered = df[df['relevant'] == 1].copy().reset_index(drop = True)
filtered.shape

### Tokenize the text and clean the tokens

stopwords = set(nltk.corpus.stopwords.words('english'))
lemma = WordNetLemmatizer()

%%time
filtered['Tokens'] = filtered['main_text'].apply(lambda x: nltk.tokenize.word_tokenize(x))

def clean_tokens(token_ls):
    
    # Remove single-character tokens (mostly punctuation)
    clean_tokens = [word for word in token_ls if len(word) > 1]

    # Remove numbers
    clean_tokens = [word for word in clean_tokens if not word.isnumeric()]

    # Remove punctuation
    clean_tokens = [word for word in clean_tokens if word.isalpha()]

    # Remove stopwords
    clean_tokens = [word for word in clean_tokens if word not in stopwords]
    
    clean_tokens = [lemma.lemmatize(word) for word in clean_tokens]
    
    return clean_tokens

%%time
filtered['cleaned_tokens'] = filtered['Tokens'].apply(clean_tokens)

table = pa.Table.from_pandas(filtered)
pq.write_table(table, './filtered_data.parquet')
