==> code/02_Topic Modeling.ipynb <==
#!pip install pandarallel

from pandarallel import pandarallel
import multiprocessing

num_processors = multiprocessing.cpu_count()
print(f'Available CPUs: {num_processors}')

pandarallel.initialize(nb_workers=num_processors-1, use_memory_fs=False, progress_bar = True)

import pyarrow as pa
import pyarrow.parquet as pq
import pandas as pd

import warnings

# warnings.simplefilter('once')
warnings.simplefilter('ignore')
# warnings.filterwarnings("ignore", category=FutureWarning)
# warnings.filterwarnings("ignore", category=DeprecationWarning)
# warnings.filterwarnings(action='ignore', category=UserWarning, module='gensim')

%%time
df = pd.read_parquet("filtered_data.parquet")

df.head()

df.shape

## LDA with gensim

!pip install textblob
!pip install nltk
!pip install gensim
!pip install pyLDAvis

import time
import math
import re
from textblob import TextBlob


import nltk as nltk
from nltk.corpus import stopwords
from nltk.stem.wordnet import WordNetLemmatizer

import string


import gensim
from gensim import corpora, models
from gensim.models.ldamulticore import LdaMulticore
from gensim.utils import simple_preprocess
from gensim.models import CoherenceModel
from nltk.stem.wordnet import WordNetLemmatizer

# import pyLDAvis.gensim
import pyLDAvis
import pyLDAvis.gensim_models as gensimvis
pyLDAvis.enable_notebook()

import multiprocessing

num_processors = multiprocessing.cpu_count()
num_processors
workers=num_processors-1

from tqdm import tqdm

# define helper function
def dict_doc_term(text_df):
    
    doc_clean = text_df.to_list()
    
    # Creating the term dictionary of our courpus, where every unique term is assigned an index. 
    dictionary = corpora.Dictionary(doc_clean)

    # Converting list of documents (corpus) into Document Term Matrix using dictionary prepared above.
    doc_term_matrix = [dictionary.doc2bow(doc) for doc in doc_clean]
    
    return doc_clean, dictionary, doc_term_matrix


def n_topic(doc_clean, dictionary, doc_term_matrix, topic_ls):
    
    coherence_score = []
    
    for n in tqdm(topic_ls):

        lda_model = LdaMulticore(corpus = doc_term_matrix,
                   id2word = dictionary,
                   num_topics = n,
                   random_state = 100,
                   passes = 10,
                   alpha = 'symmetric',
                   eta = 'auto',
                   workers = workers)
        
        coherence_model_lda = CoherenceModel(model = lda_model, texts = doc_clean, dictionary = dictionary, coherence = 'c_v')
        coherence_lda = coherence_model_lda.get_coherence()
        coherence_score.append(coherence_lda)
        
        print(f'\nWith {n} topics:')
        print(*lda_model.print_topics(num_topics = n, num_words = 10), sep='\n')
        
    return pd.DataFrame({"n": topic_ls, "coherence_score": coherence_score})

%%time
df_cleaned, df_dict, df_matrix = dict_doc_term(df['cleaned_tokens'])

#%%time
df_score = n_topic(df_cleaned, df_dict, df_matrix, [16])

df_score

n_topics = [4, 6, 8, 10, 12, 14, 16]
coherence_scores = [0.525972, 0.491855, 0.483164, 0.481007, 0.490952, 0.477404, 0.463993]
time = ["10:30", "13:46", "16:37", "26:21", "36:22", "48:30", "1:03:06"]

gensim_res = pd.DataFrame({"n_topic": n_topics, "coherence_score": coherence_scores, "time": time})
gensim_res

## LDA with K-Train

#!pip install bokeh 
#!pip install ktrain 

import bokeh
import ktrain
print('Bokeh Version: ' + bokeh.__version__)
print('Ktrain Version: ' + ktrain.__version__)

texts = df['main_text'].tolist()

%%time
tm = ktrain.text.get_topic_model(
    texts=texts, 
    n_topics=20, 
    n_features=10000, 
    min_df=5, 
    max_df=0.5, 
    stop_words='english', 
    model_type='lda', 
    lda_max_iter=5, 
    verbose=1)

%%time
tm.build(texts, threshold=0.25)

tm.print_topics(show_counts=True)

%%time
topic_doc = []

for i in range(20):
    topic_doc.append([text['text'] for text in tm.get_docs(topic_ids=[i], rank=True)])

def assign_topic(text):
    
    for idx, topic in enumerate(topic_doc):
        if text in topic:
            return str(idx)

df['ktrain_topic'] = df['main_text'].parallel_apply(assign_topic)

df.head()

df['ktrain_topic'].value_counts()

import pyarrow as pa
import pyarrow.parquet as pq
table = pa.Table.from_pandas(df)
pq.write_table(table, './data_topic.parquet')

#doc_topics=tm.get_doctopics()
#tm.visualize_documents(doc_topics=tm.get_doctopics())

## BERTopic

#!pip install bertopic

from bertopic import BERTopic

### Cleaned_tokens

%%time
articles = df['cleaned_tokens'].tolist()
text = [' '.join(tokens) for tokens in articles]

%%time
# Create an instance of BERTopic
#topic_model = BERTopic(embedding_model = "bert-base-nli-mean-tokens",  min_topic_size = 150)
topic_model = BERTopic(embedding_model = "bert-base-nli-mean-tokens",  min_topic_size = 50)
# Fit the BERTopic model on your data
topics, _ = topic_model.fit_transform(text)

# Get the most frequent topics
top_topics = topic_model.get_topic_freq()

# Print the top N topics
N = 20

for topic in top_topics.head(N)['Topic']:
    topic_words = [word[0] for word in topic_model.get_topic(topic)]
    print(f"Topic {topic}: {', '.join(topic_words)}")

# Save model
topic_model.save('bertopic2')

res = topic_model.get_document_info(text)
res['Name'].value_counts()

### Title 

#!pip install nltk

import nltk
#nltk.download('stopwords')
from nltk.corpus import stopwords

stop_words = stopwords.words('english')

title_ls = df['title'].tolist()
#convert to lowercase
title_ls = [title.lower() for title in title_ls]
#remove stopwords
cleaned_title = []
for title in title_ls: 
    tokens = title.split(' ')
    cleaned_tokens = []
    for token in tokens:
        if token not in stop_words:
            cleaned_tokens.append(token)
            
    cleaned_title.append(' '.join(cleaned_tokens))

%%time
# Create an instance of BERTopic
title_topic_model = BERTopic(embedding_model = "bert-base-nli-mean-tokens",  min_topic_size = 150)

# Fit the BERTopic model on your data
title_topics, _ = title_topic_model.fit_transform(cleaned_title)

# Get the most frequent topics
title_top_topics = title_topic_model.get_topic_freq()

# Print the top N topics
N = 20

for topic in title_top_topics.head(N)['Topic']:
    topic_words = [word[0] for word in title_topic_model.get_topic(topic)]
    print(f"Topic {topic}: {', '.join(topic_words)}")

# Save model
title_topic_model.save('title_bertopic')

res = title_topic_model.get_document_info(cleaned_title)
res['Name'].value_counts()

### Identify topics

import re

saved_model = BERTopic.load('bertopic2')

%%time
# Get the most frequent topics
new_top_topics = saved_model.get_topic_freq()
# Print the top N topics
N = 40

for topic in new_top_topics.head(N)['Topic']:
    topic_words = [word[0] for word in saved_model.get_topic(topic)]
    print(f"Topic {topic}: {', '.join(topic_words)}")

#assign industry tag to each topic
industry_name = ['Others']

for i in range(len(new_top_topics['Topic'])-1):
    
    sample = [word[0] for word in saved_model.get_topic(i)]
    
    if re.search(r"crypto(currency)?|blockchain|finance|btc|bank(ing)?|loan|fintech", ' '.join(sample), flags=re.I): industry_name.append('Finance')
    elif re.search(r"patient|health(care)?|cancer|clinical|drug|medicine|genomics", ' '.join(sample), flags=re.I): industry_name.append('Healthcare&Biotech')
    elif re.search(r"student|classroom|education|coursera|essay|school|academic", ' '.join(sample), flags=re.I): industry_name.append('Education')
    elif re.search(r"music|song|photoshop|painting|artist|photography|film", ' '.join(sample), flags=re.I): industry_name.append('Media Creation')
    elif re.search(r"automative|vehicle|tesla|elon|car|energy|automobile", ' '.join(sample), flags=re.I): industry_name.append('Automotive')
    elif re.search(r"telecom|telecast|5g|IoT|mobile|satellite|fiber optics", ' '.join(sample), flags=re.I): industry_name.append('Telecommunication')
    elif re.search(r"agriculture|farm(ing)?|agribotix|farmer|livestock|pest|soil", ' '.join(sample), flags=re.I): industry_name.append('Agriculture')
    elif re.search(r"writer|publication(s)?|book(s)?|publishing|magazine|author|newspapers", ' '.join(sample), flags=re.I): industry_name.append('Publishing')
    elif re.search(r"google|ibm|microsoft|facebook|chip(s)|meta|chatgpt", ' '.join(sample), flags=re.I): industry_name.append('Technology')
    elif re.search(r"military|defense|aircraft|space|cybersecurity|drone|army", ' '.join(sample), flags=re.I): industry_name.append('Military & Defense')
    else: industry_name.append('Others')
    

topic_cnt = saved_model.get_document_info(text).groupby('Name').count().sort_values(by = 'Document', ascending = False).reset_index()
topic_cnt['industry'] = industry_name
topic_cnt.head()

# map the industry label to the dataframe
industry_label = {'Key': topic_cnt['Name'].tolist(),
                  'Value': topic_cnt['industry'].tolist()}

industry_label_df = pd.DataFrame(industry_label)

result_dict = industry_label_df.set_index('Key')['Value'].to_dict()

doc_info = saved_model.get_document_info(text)
doc_info.head()

%%time
doc_info['Industry'] = doc_info['Name'].apply(lambda x: result_dict[x])

doc_info.head()

doc_info.groupby('Industry').count().sort_values(by = 'Document', ascending = False)

doc_info.to_csv('industry_info.csv')
