==> code/03_Sentiment Analysis.ipynb <==
#!pip install vaderSentiment
#!pip install pandarallel

import pandas as pd
import numpy as np
import os
import json
import glob
import datetime
import time
import pytz
#Load the SentimentIntensityAnalyzer object from the VADER package
from vaderSentiment.vaderSentiment import SentimentIntensityAnalyzer

from pandarallel import pandarallel
import multiprocessing

num_processors = multiprocessing.cpu_count()
print(f'Available CPUs: {num_processors}')

pandarallel.initialize(nb_workers=num_processors-1, use_memory_fs=False, progress_bar = True)

%%time
df = pd.read_parquet("data_topic.parquet")

df.head()

df.shape

## VADER

#Create a handle to the SentimentIntensityAnalyzer object
analyzer = SentimentIntensityAnalyzer()

def vader_score(text):
    score = analyzer.polarity_scores(text)
    return score['compound']

%%time
df['vader_score'] = df['main_text'].parallel_apply(vader_score)

df['vader_sentiment'] = np.where(df['vader_score'].astype(np.float) > 0, 'Positive', 'Negative')
df['vader_sentiment'] = np.where(df['vader_score'].astype(np.float) == 0, 'Neutral', df['vader_sentiment'])

df.head()

import pyarrow as pa
import pyarrow.parquet as pq
table = pa.Table.from_pandas(df)
pq.write_table(table, './data_vader_score.parquet')

## SVM trained on Yelp data

import pickle
import pandas as pd

%%time
df = pd.read_parquet("data_vader_score.parquet")

df.head()

with open('vect.pkl', 'rb') as f, open('svm_sentiment.pkl', 'rb') as model:
    vect = pickle.load(f)
    svm = pickle.load(model)

text = [' '.join(article) for article in df['cleaned_tokens'].tolist()]

%%time
y_pred = svm.predict(vect.transform(text))

df['svm_sentiment'] = y_pred
df.head()

import pyarrow as pa
import pyarrow.parquet as pq
table = pa.Table.from_pandas(df)
pq.write_table(table, './data_sentiment_topic.parquet')

## Customized Vader

new = pd.read_parquet('data_sentiment_topic.parquet')
new.head()

#Create a handle to the SentimentIntensityAnalyzer object
analyzer2 = SentimentIntensityAnalyzer()

#Update the analyzer with customized word list
pos = ['milestone', 'opportunity', 'growth', 'innovation', 'efficiency', 'productivity', 'competitiveness', 'prospects', 
       'sustainability', 'development', 'progress', 'transformation', 'digitization', 'smart', 'wisdom', 
       'optimization', 'data-driven', 'prediction', 'accuracy', 'high-quality', 'high-efficiency', 'cost-saving', 'safety', 
       'reliability', 'sustainability', 'future-oriented', 'leading', 'leader', 'advantage', 'forward-looking', 
       'exploration', 'creation', 'experimentation', 'discovery', 'breakthrough']

neg = ['unemployment', 'uncertainty', 'contraction', 'reduction', 'downsizing', 'bottleneck', 'challenge', 'obstacle', 'risk', 
       'insecurity', 'instability', 'unreliability', 'opacity', 'restriction', 'limitation', 'regression', 'stagnation', 
       'backwardness', 'failure', 'error', 'defect', 'problem', 'difficulty', 'negative', 'pessimistic', 'worry', 'fear', 
       'inadequacy', 'pressure', 'struggle', 'plight', 'crisis', 'frustration', 'disappointment', 'injury', 'hopelessness', 
       'pessimism', 'downward trend', 'stubbornness', 'conservatism', 'resistance', 'constraint', 'stagnation', 'inflexibility', 
       'outdated', 'slow', 'low-efficiency', 'low-quality', 'high-cost', 'adverse impact', 'loss', 'high risk', 'heavy burden', 
       'drag', 'high challenge', 'detrimental to human beings', 'opposition', 'confrontation']

analyzer2.lexicon.update({word: 1.0 for word in pos})
analyzer2.lexicon.update({word: -1.0 for word in neg})

def vader_score2(text):
    score = analyzer2.polarity_scores(text)
    return score['compound']

%%time
new['vader_score2'] = new['main_text'].parallel_apply(vader_score2)

new['vader_sentiment2'] = np.where(new['vader_score2'].astype(np.float) > 0, 'Positive', 'Negative')
new['vader_sentiment2'] = np.where(new['vader_score2'].astype(np.float) == 0, 'Neutral', new['vader_sentiment'])

new.head()

## PatternAnalyzer from textblob

!pip install textblob

from textblob import TextBlob
from textblob.sentiments import PatternAnalyzer

patternanalyzer = PatternAnalyzer()

def textblob_sent(text):
    score = TextBlob(text, analyzer = patternanalyzer).sentiment.polarity
    
    if score > 0: return 'Positive'
    elif score < 0: return 'Negative'
    else: return 'Neutral'

%%time
new['textblob_sent'] = new['main_text'].parallel_apply(textblob_sent)

new.head()

import pyarrow as pa
import pyarrow.parquet as pq
table = pa.Table.from_pandas(new)
pq.write_table(table, './data_sentiment_topic.parquet')

## Transformer trained on financial news

import pickle 
import pandas as pd
from tqdm import tqdm

from transformers import AutoTokenizer, AutoModelForSequenceClassification
import torch 

tokenizer = AutoTokenizer.from_pretrained("mrm8488/distilroberta-finetuned-financial-news-sentiment-analysis")

model = AutoModelForSequenceClassification.from_pretrained("mrm8488/distilroberta-finetuned-financial-news-sentiment-analysis")

%%time
with open('data_res.pkl', 'rb') as f:
    df = pickle.load(f)

def get_sentiment(text):
    
    inputs = tokenizer(text, padding = True, truncation = True, return_tensors = 'pt')
    logits = model(**inputs).logits
    
    predicted_class_id = logits.argmax().item()
    label = model.config.id2label[predicted_class_id]

    return label

%%time
bert_sentiment = []
for text in tqdm(df['main_text'].values):
    bert_sentiment.append(get_sentiment(text))

df['bert_sentiment'] = bert_sentiment

%%time
import pickle

with open('data_full.pkl', 'wb') as f:
    pickle.dump(df, f)

## result comparison

new['textblob_sent'].value_counts()

new['vader_sentiment'].value_counts()

new['vader_sentiment2'].value_counts()

new['svm_sentiment'].value_counts()
