==> code/04_NER.ipynb <==
!pip install nltk
!pip install --user spacy
!pip install pandarallel
!pip install --user https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.0.0/en_core_web_sm-3.0.0.tar.gz

import pandas as pd
import nltk as nltk
import nltk.corpus  
from nltk.text import Text
import matplotlib.pyplot as plt
import seaborn as sns
import re 
import spacy 
from spacy import displacy
import warnings
warnings.filterwarnings("ignore")

#pd.set_option('display.max_rows', 100)
#pd.set_option('display.max_columns', None)
#pd.set_option('display.max_colwidth', 500)

import multiprocessing
import pandarallel
from pandarallel import pandarallel

# Set up pandarallel
num_processors = multiprocessing.cpu_count()
pandarallel.initialize(nb_workers=num_processors-1, use_memory_fs=False, progress_bar = True)

nlp = spacy.load("en_core_web_sm")
# Checking active pipeline components
nlp.pipe_names

%%time
df = pd.read_parquet("data_sentiment_topic.parquet")
df.head()

## NER using SpaCy without sentence segmentation

# define helper function
def text_ner(text):
    
    entities = []
    labels = []
    
    doc = nlp(text)

    for ent in doc.ents:
        if ent.label_ in ['ORG', 'GPE', 'LOC', 'PERSON']:
            entities.append(ent.text)
            labels.append(ent.label_)
            
    entities_labels_df = pd.DataFrame({"Entities": entities, "Labels": labels})
    
    return entities_labels_df

df['main_text'].values[13]

%%time
df['NER'] = df['main_text'].parallel_apply(text_ner)

df.head()

import pickle

with open('data_res.pkl', 'wb') as f:
    pickle.dump(df, f)

## Extract NER from each sentence and calculate the sentiment score for these entities
- calculate the sentence sentiment with customized vader

!pip install vaderSentiment

from vaderSentiment.vaderSentiment import SentimentIntensityAnalyzer
import pickle
import pandas as pd

%%time
with open('data_full.pkl', 'rb') as f:
    df = pickle.load(f)

analyzer = SentimentIntensityAnalyzer()
#Update the analyzer with customized word list
pos = ['milestone', 'opportunity', 'growth', 'innovation', 'efficiency', 'productivity', 'competitiveness', 'prospects', 
       'sustainability', 'development', 'progress', 'transformation', 'digitization', 'smart', 'wisdom', 
       'optimization', 'data-driven', 'prediction', 'accuracy', 'high-quality', 'high-efficiency', 'cost-saving', 'safety', 
       'reliability', 'sustainability', 'future-oriented', 'leading', 'leader', 'advantage', 'forward-looking', 
       'exploration', 'creation', 'experimentation', 'discovery', 'breakthrough']

neg = ['unemployment', 'uncertainty', 'contraction', 'reduction', 'downsizing', 'bottleneck', 'challenge', 'obstacle', 'risk', 
       'insecurity', 'instability', 'unreliability', 'opacity', 'restriction', 'limitation', 'regression', 'stagnation', 
       'backwardness', 'failure', 'error', 'defect', 'problem', 'difficulty', 'negative', 'pessimistic', 'worry', 'fear', 
       'inadequacy', 'pressure', 'struggle', 'plight', 'crisis', 'frustration', 'disappointment', 'injury', 'hopelessness', 
       'pessimism', 'downward trend', 'stubbornness', 'conservatism', 'resistance', 'constraint', 'stagnation', 'inflexibility', 
       'outdated', 'slow', 'low-efficiency', 'low-quality', 'high-cost', 'adverse impact', 'loss', 'high risk', 'heavy burden', 
       'drag', 'high challenge', 'detrimental to human beings', 'opposition', 'confrontation']

analyzer.lexicon.update({word: 1.0 for word in pos})
analyzer.lexicon.update({word: -1.0 for word in neg})

analyzer.polarity_scores(df['main_text'].values[0])

def vader_sent(text):
    score = analyzer.polarity_scores(text)
    return score['compound']

# define helper function
def sentence_ner(text):

    entities = []
    labels = []
    score_ls = []
    sentiment = []

    for sent in nltk.sent_tokenize(text):
        doc = nlp(sent)
        for ent in doc.ents:
            if ent.label_ in ['ORG', 'GPE', 'LOC', 'PERSON', 'PRODUCT']:
                entities.append(ent.text)
                labels.append(ent.label_)
                score = vader_sent(sent)
                score_ls.append(score)
                if score > 0: sentiment.append('Positive')
                elif score < 0: sentiment.append('Negative')
                else: sentiment.append('Neutral')
            
    entities_sent_df = pd.DataFrame({"Entities": entities, "Labels": labels, "Score": score_ls, "Sentiment": sentiment})
    
    return entities_sent_df

df.head()

%%time
df['NER_sent'] = df['main_text'].parallel_apply(sentence_ner)

#ner = pd.concat(df['NER'].tolist())
sent_ner = pd.concat(df['NER_sent'].tolist())

ner.groupby('Entities').count().sort_values(by = 'Labels', ascending = False).head(10)

sent_ner.groupby('Entities').count().sort_values(by = 'Labels', ascending = False).head(10)

df.head()

%%time
import pickle

with open('data_full2.pkl', 'wb') as f:
    pickle.dump(df, f)

%%time
with open('data_full2.pkl', 'rb') as f:
    res = pickle.load(f)

res.head()
